moving structs to bean
